ADD dynamic inputs
